tidying up

+++ FTMContextNet/Application/Services/DeleteImportService.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using FTMContextNet.Data.Repositories;
 using FTMContextNet.Data.Repositories.GedImports;
 using FTMContextNet.Domain.Auth;
+using FTMContextNet.Domain.Entities.NonPersistent;
 using LoggingLib;
 
 namespace FTMContextNet.Application.Services
@@ -31,29 +32,46 @@ namespace FTMContextNet.Application.Services
             _auth = auth;
         }
 
-        public void Execute()
+        public APIResult Execute()
         {
+            if (_auth.GetUser() == -1)
+            {
+                return new APIResult
+                {
+                    ApiResultType = APIResultType.Unauthorized
+                };
+            }
+
             var importId = _persistedImportCacheRepository.GetCurrentImportId();
         
             //personview
             _persistedCacheRepository.DeletePersons(importId);
+            _ilog.WriteLine("Deleting persons for import id: " + importId);
             //marriages
             _persistedCacheRepository.DeleteMarriages(importId);
-           
-           
+            _ilog.WriteLine("Deleting marriages for import id: " + importId);
             //dupeentries
             _persistedCacheRepository.DeleteDupes(importId);
+            _ilog.WriteLine("Deleting dupes for import id: " + importId);
             //personorigins
-            _persistedCacheRepository.DeleteOrigins(_auth.GetUser());
+            _persistedCacheRepository.DeleteOrigins(importId);
+            _ilog.WriteLine("Deleting origins for import id: " + importId);
             //treerecordmapgroups
             _persistedCacheRepository.DeleteRecordMapGroups(importId);
+            _ilog.WriteLine("Deleting record map groups for import id: " + importId);
             //treegroups
             _persistedCacheRepository.DeleteTreeGroups(importId);
+            _ilog.WriteLine("Deleting tree groups for import id: " + importId);
             //treerecords
             _persistedCacheRepository.DeleteTreeRecords(importId);
-
+            _ilog.WriteLine("Deleting tree records for import id: " + importId);
             //import
             _persistedImportCacheRepository.DeleteImport(importId);
+
+            return new APIResult
+            {
+                ApiResultType = APIResultType.Success
+            };
         }
     }
 }
